requirements.txt
fastapi
uvicorn
polars
apscheduler
requests
bs4
libarchive-c
concurrent-log-handler
